src/requirements.txt
aiofiles==23.2.1
aioredis==2.0.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
click==8.1.7
dnspython==2.4.2
exceptiongroup==1.2.0
fastapi==0.105.0
flake8==6.1.0
h11==0.14.0
idna==3.6
install==1.3.5
isort==5.13.1
loguru==0.7.2
mccabe==0.7.0
motor==3.3.2
pycodestyle==2.11.1
pydantic==1.10.13
pydantic_core==2.14.5
pyflakes==3.1.0
pymongo==4.6.1
python-dotenv==1.0.0
python-multipart==0.0.6
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.9.0
uvicorn==0.24.0.post1
